refactor classifiers class names
