+++ freqtrade/freqai/prediction_models/PyTorchClassifier.py
